Purge unused code
