+++ cellrank/pl/__init__.py
@@ -1,8 +1,7 @@
 # -*- coding: utf-8 -*-
 from cellrank.pl._graph import graph
-from cellrank.pl._utils import composition
 from cellrank.pl._heatmap import heatmap
 from cellrank.pl._lineages import lineages, lineage_drivers
 from cellrank.pl._gene_trend import gene_trends
-from cellrank.pl._cluster_fates import cluster_fates, similarity_plot
+from cellrank.pl._cluster_fates import cluster_fates
 from cellrank.pl._cluster_lineages import cluster_lineage
