+++ cellrank/tl/__init__.py
@@ -1,10 +1,7 @@
 # -*- coding: utf-8 -*-
 import cellrank.tl.kernels
 import cellrank.tl.estimators
-from cellrank.tl._utils import partition, cyto_trace
 from cellrank.tl._lineage import Lineage
-from cellrank.tl._lineages import lineages
-from cellrank.tl._lineage_drivers import lineage_drivers
-from cellrank.tl._permutation_test import _permutation_test
+from cellrank.tl._lineages import lineages, lineage_drivers
 from cellrank.tl._transition_matrix import transition_matrix
 from cellrank.tl._initial_or_terminal_states import initial_states, terminal_states
