+++ cellrank/ul/_utils.py
@@ -6,7 +6,7 @@ from functools import wraps, update_wrapper
 from multiprocessing import cpu_count
 
 import numpy as np
-from scipy.sparse import spmatrix
+from scipy.sparse import issparse, spmatrix
 
 from cellrank import logging as logg
 
@@ -224,3 +224,12 @@ def valuedispatch(func):
     update_wrapper(wrapper, func)
 
     return wrapper
+
+
+def _densify_squeeze(x: Union[spmatrix, np.ndarray], dtype=np.float32) -> np.ndarray:
+    if issparse(x):
+        x = x.toarray()
+    if x.ndim == 2 and x.shape[1] == 1:
+        x = np.squeeze(x, axis=1)
+
+    return x[:].astype(dtype)
