+++ tests/test_tools.py
@@ -1,58 +1,14 @@
 # -*- coding: utf-8 -*-
 import pytest
-from _helpers import create_model
 
 from anndata import AnnData
 
 import pandas as pd
 
 import cellrank as cr
+import cellrank.tl._lineages
 from cellrank.tl.kernels import Kernel
 from cellrank.tl._constants import AbsProbKey, FinalStatesKey, _probs
-from cellrank.tl._lineage_drivers import _gene_importance
-
-
-class TestGeneImportance:
-    def test_invalid_n_perms(self, adata_cflare):
-        model = create_model(adata_cflare)
-        with pytest.raises(ValueError):
-            _ = _gene_importance(
-                adata_cflare, model, adata_cflare.var_names[:10], "0", n_perms=-1
-            )
-
-    def test_no_perms(self, adata_cflare):
-        model = create_model(adata_cflare)
-        res = _gene_importance(
-            adata_cflare, model, adata_cflare.var_names[:10], "0", n_perms=None
-        )
-
-        assert isinstance(res, pd.DataFrame)
-        assert res.shape == (10, 1)
-        assert set(res.index) == set(adata_cflare.var_names[:10])
-        assert "importance" in res.columns
-
-    def test_perms_no_fdr_correction(self, adata_cflare):
-        model = create_model(adata_cflare)
-        res = _gene_importance(
-            adata_cflare,
-            model,
-            adata_cflare.var_names,
-            "0",
-            n_perms=10,
-            fdr_correction=None,
-        )
-
-        assert "pval" in res.columns
-        assert "qval" not in res.columns
-
-    def test_perms_fdr_correction(self, adata_cflare):
-        model = create_model(adata_cflare)
-        res = _gene_importance(
-            adata_cflare, model, adata_cflare.var_names, "0", n_perms=10
-        )
-
-        assert "pval" in res.columns
-        assert "qval" in res.columns
 
 
 class TestLineages:
@@ -75,11 +31,11 @@ class TestLineages:
 class TestLineageDrivers:
     def test_no_abs_probs(self, adata: AnnData):
         with pytest.raises(RuntimeError):
-            cr.tl.lineage_drivers(adata)
+            cellrank.tl._lineages.lineage_drivers(adata)
 
     def test_normal_run(self, adata_cflare):
         cr.tl.lineages(adata_cflare)
-        cr.tl.lineage_drivers(adata_cflare, use_raw=False)
+        cellrank.tl._lineages.lineage_drivers(adata_cflare, use_raw=False)
 
         for name in adata_cflare.obsm[AbsProbKey.FORWARD.s].names:
             assert f"to {name}" in adata_cflare.var
@@ -87,14 +43,16 @@ class TestLineageDrivers:
     def test_normal_run_raw(self, adata_cflare):
         adata_cflare.raw = adata_cflare.copy()
         cr.tl.lineages(adata_cflare)
-        cr.tl.lineage_drivers(adata_cflare, use_raw=True)
+        cellrank.tl._lineages.lineage_drivers(adata_cflare, use_raw=True)
 
         for name in adata_cflare.obsm[AbsProbKey.FORWARD.s].names:
             assert f"to {name}" in adata_cflare.raw.var
 
     def test_return_drivers(self, adata_cflare):
         cr.tl.lineages(adata_cflare)
-        res = cr.tl.lineage_drivers(adata_cflare, return_drivers=True, use_raw=False)
+        res = cellrank.tl._lineages.lineage_drivers(
+            adata_cflare, return_drivers=True, use_raw=False
+        )
 
         assert isinstance(res, pd.DataFrame)
         assert res.shape[0] == adata_cflare.n_vars
@@ -184,23 +142,3 @@ class TestTransitionMatrix:
         kernel_add = cr.tl.transition_matrix(adata, backward=True)
 
         assert isinstance(kernel_add, Kernel)
-
-
-class TestCytoTrace:
-    def test_wrong_layer(self, adata: AnnData):
-        with pytest.raises(KeyError):
-            cr.tl.cyto_trace(adata, layer="foo")
-
-    def test_normal_run(self, adata: AnnData):
-        cr.tl.cyto_trace(adata)
-
-        assert "gcs" in adata.obs.keys()
-        assert "gene_corr" in adata.var.keys()
-        assert "correlates" in adata.var.keys()
-
-    def test_normal_run_copy(self, adata: AnnData):
-        adata = cr.tl.cyto_trace(adata, inplace=False)
-
-        assert "gcs" in adata.obs.keys()
-        assert "gene_corr" in adata.var.keys()
-        assert "correlates" in adata.var.keys()
